[CSL-1081] Move genesis/main block/header constructors into Block.Core
