+++ src/Pos/Block/Core/Main/Misc.hs
@@ -4,15 +4,19 @@
 -- | Miscellaneous instances, etc. Related to the main blockchain of course.
 
 module Pos.Block.Core.Main.Misc
-       (
+       ( mkMainBlock
+       , recreateMainBlock
+       , mkMainHeader
        ) where
 
 import           Universum
 
+import           Control.Monad.Except       (MonadError (throwError))
 import qualified Data.Text.Buildable        as Buildable
 import           Formatting                 (bprint, build, int, stext, (%))
 import           Serokell.Util              (Color (Magenta), colorize, listJson)
 
+import           Pos.Binary.Block.Core      ()
 import           Pos.Block.Core.Main.Chain  (Body (..), ConsensusData (..))
 import           Pos.Block.Core.Main.Lens   (mainBlockBlockVersion, mainBlockDifficulty,
                                              mainBlockSlot, mainBlockSlot,
@@ -22,17 +26,23 @@ import           Pos.Block.Core.Main.Lens   (mainBlockBlockVersion, mainBlockDif
                                              mainHeaderSoftwareVersion, mbTxs,
                                              mcdDifficulty, mehBlockVersion,
                                              mehSoftwareVersion)
-import           Pos.Block.Core.Main.Types  (MainBlock, MainBlockHeader, MainBlockchain,
-                                             MainExtraHeaderData)
-import           Pos.Block.Core.Union.Types (BiHeader, BiSsc, blockHeaderHash)
-import           Pos.Core                   (EpochOrSlot (..), GenericBlock (..),
-                                             GenericBlockHeader (..),
+import           Pos.Block.Core.Main.Types  (BlockSignature (..), MainBlock,
+                                             MainBlockHeader, MainBlockchain,
+                                             MainExtraBodyData (..), MainExtraHeaderData,
+                                             MainToSign (..))
+import           Pos.Block.Core.Union.Types (BiHeader, BiSsc, BlockHeader,
+                                             blockHeaderHash)
+import           Pos.Core                   (Blockchain (..), EpochOrSlot (..),
+                                             GenericBlock (..), GenericBlockHeader (..),
                                              HasBlockVersion (..), HasDifficulty (..),
                                              HasEpochIndex (..), HasEpochOrSlot (..),
                                              HasHeaderHash (..), HasSoftwareVersion (..),
                                              HeaderHash, IsHeader, IsMainHeader (..),
+                                             ProxySKEither, SlotId, mkGenericHeader,
                                              slotIdF)
-import           Pos.Crypto                 (hashHexF)
+import           Pos.Crypto                 (ProxySecretKey (..), SecretKey, SignTag (..),
+                                             hashHexF, proxySign, sign, toPublic)
+import           Pos.Ssc.Class.Helpers      (SscHelpersClass (..))
 
 instance BiSsc ssc => Buildable (MainBlockHeader ssc) where
     build gbh@GenericBlockHeader {..} =
@@ -132,3 +142,76 @@ instance BiHeader ssc => IsHeader (MainBlockHeader ssc)
 instance BiHeader ssc => IsMainHeader (MainBlockHeader ssc) where
     headerSlotL = mainHeaderSlot
     headerLeaderKeyL = mainHeaderLeaderKey
+
+----------------------------------------------------------------------------
+-- Smart constructors
+----------------------------------------------------------------------------
+
+type SanityConstraint ssc
+     = ( BiSsc ssc
+       , SscHelpersClass ssc
+       , HasDifficulty $ BlockHeader ssc
+       , HasHeaderHash $ BlockHeader ssc
+       )
+
+-- | Smart constructor for 'MainBlockHeader'.
+mkMainHeader
+    :: (SanityConstraint ssc)
+    => Maybe (BlockHeader ssc)
+    -> SlotId
+    -> SecretKey
+    -> Maybe ProxySKEither
+    -> Body (MainBlockchain ssc)
+    -> MainExtraHeaderData
+    -> MainBlockHeader ssc
+mkMainHeader prevHeader slotId sk pSk body extra =
+        mkGenericHeader prevHeader body consensus extra
+  where
+    difficulty = maybe 0 (succ . view difficultyL) prevHeader
+    makeSignature toSign (Left psk) = BlockPSignatureLight $
+        proxySign SignMainBlockLight sk psk toSign
+    makeSignature toSign (Right psk) = BlockPSignatureHeavy $
+        proxySign SignMainBlockHeavy sk psk toSign
+    signature prevHash proof =
+        let toSign = MainToSign prevHash proof slotId difficulty extra
+        in maybe (BlockSignature $ sign SignMainBlock sk toSign)
+                 (makeSignature toSign)
+             pSk
+    consensus prevHash proof =
+        MainConsensusData
+        { _mcdSlot = slotId
+        , _mcdLeaderKey = maybe (toPublic sk) (either pskIssuerPk pskIssuerPk) pSk
+        , _mcdDifficulty = difficulty
+        , _mcdSignature = signature prevHash proof
+        }
+
+-- | Smart constructor for 'MainBlock'. Uses 'mkMainHeader'. It
+-- verifies consistency of given data and may fail.
+mkMainBlock
+    :: (SanityConstraint ssc, MonadError Text m)
+    => Maybe (BlockHeader ssc)
+    -> SlotId
+    -> SecretKey
+    -> Maybe ProxySKEither
+    -> Body (MainBlockchain ssc)
+    -> MainExtraHeaderData
+    -> MainExtraBodyData
+    -> m (MainBlock ssc)
+mkMainBlock prevHeader slotId sk proxyInfo body extraH extraB =
+    recreateMainBlock
+        (mkMainHeader prevHeader slotId sk proxyInfo body extraH)
+        body
+        extraB
+
+-- | Recreate 'MainBlock' from existing data. It also verifies
+-- consistency of this data and may fail.
+recreateMainBlock
+    :: (SanityConstraint ssc, MonadError Text m)
+    => MainBlockHeader ssc
+    -> Body (MainBlockchain ssc)
+    -> MainExtraBodyData
+    -> m (MainBlock ssc)
+recreateMainBlock _gbHeader _gbBody _gbExtra = do
+    let gb = GenericBlock{..}
+    whenLeft (verifyBBlock gb) throwError
+    pure gb
